+++ src/Structures/MenuDocsClient.js
@@ -1,7 +1,9 @@
 const { Client, Collection, Intents } = require('discord.js');
 const Util = require('./Util.js');
 
+//Diese Klasse weißt dem Discord-Bot Standardparameter zu und kann ihn starten.
 module.exports = class MenuDocsClient extends Client {
+    //Initialisere und weise dem Bot seine Daten zu.
     constructor(options = {}) {
         console.log('MenuDocsClient.js called >> running constructor');
         super({
@@ -26,9 +28,9 @@ module.exports = class MenuDocsClient extends Client {
         this.owners = options.DiscordAPI.owners;
     }
 
-
+    //Überprüfe, ob die Optionen, die in "config.json" angegeben wurden, auch gültig sind
     validate(options) {
-        console.log('MenuDocsClient.js called >> validating options');
+        console.log('Validating \"config.json\"...');
         if(typeof options !== 'object') throw new TypeError('Options should be type of Object.');
 
         if(!options.DiscordAPI.token) throw new Error('You must pass the token for the client.');
@@ -37,14 +39,15 @@ module.exports = class MenuDocsClient extends Client {
         if(!options.DiscordAPI.prefix) throw new Error('You must pass a prefix for the client.');
         if(typeof options.DiscordAPI.prefix !== 'string') throw new TypeError('Prefix should be a type of String');
         this.prefix = options.DiscordAPI.prefix;
+        console.log(`Content of Configuration is valid.`);
     }
 
+    //Starte den Bot, in dem alle Befehle und Events geladen werden und sich der Bot in die Discord-API einloggt
     async start(token = this.token) {
-        console.log('MenuDocsClient.js called >> starting');
         this.utils.loadCommands();
         this.utils.loadEvents();
         super.login(token);
-        console.log('MenuDocsClient.js called >> looged in');
+        console.log('Connection successful.\n________________________________________________\n\n');
     }
 
 };